final v2 final_final

+++ UnitDataProcessing/CriteriaTerritorialUnitEducationPortion.h
@@ -11,7 +11,7 @@ public:
 	CriteriaTerritorialUnitEducationPortion(Education education) : education_(education)
 	{}
 	double evaluate(const TerritorialUnit& unit) {
-		double value = unit.getPeopleCount() == 0 ? 0 : (unit.getEducation(education_) / unit.getPeopleCount() * 100);
+		double value = unit.getPeopleCount() == 0 ? 0 : (static_cast<double>(unit.getEducation(education_)) / static_cast<double>(unit.getPeopleCount()) * 100);
 		return value;
 	}
 };
